Move ModalContextProvider to shared
